requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
supabase>=2.4.0
google-cloud-aiplatform>=1.45.0
python-dotenv>=1.0.0
openpyxl>=3.0.0
tabulate>=0.9.0